+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Kutumova 2024 albumin nanoparticles ([doi:10.1016/j.heliyon.2024.e30962](https://doi.org/10.1016/j.heliyon.2024.e30962)) - whole-body PBPK in male BALB/c mice with lipopolysaccharide-induced acute lung injury.
+
 - Add Kuroda 2023 cephalothin ([doi:10.1294/jes.34.111](https://doi.org/10.1294/jes.34.111)) - Thoroughbred horses given intramuscular and intravenous doses.
 
 - Add Zhang 2024 sertraline ([doi:10.1016/j.heliyon.2024.e25231](https://doi.org/10.1016/j.heliyon.2024.e25231)) - Chinese inpatients with psychiatric disorders, aged 11-79 years.